Refactor generation
Moved the `nielsen` generator into seperate folder as preparation for multiple generator bbackends.
